Refactor: fix communicator.js
add Parse hex to string feature, delete receiver part of HTML
