+++ code-review/index.ejs
@@ -36,13 +36,8 @@
 
 
   <div class="main__wrapper">
-    <div class="main__receiver">
-      <div>수신정보</div>
-      <input class="receiver--input" placeholder="미구현" type="text">
-      <button class="receiver--button">수신받기</button>
-    </div>
     <div class="main__result">
-      <div>해석결과</div>
+      <div>송수신정보</div>
       <input class="result--input" placeholder="미구현" type="text">
       <button class="result--button">해석하기</button>
     </div>
