+++ martian/public/js/main.js
@@ -1,5 +1,6 @@
 import { _ } from './util.js'
 import { drawWheel } from './wheel.js'
+import {parse, send} from './Communicator.js'
 
 const wheelData = {
     canvas: _.$('.wheel__circle'),
@@ -9,3 +10,15 @@ const wheelData = {
     smallerCircleRadius: 50
 }
 drawWheel(wheelData)
+
+const selectors = {
+    resultInput: _.$('.result--input'),
+    resultButton: _.$('.result--button'),
+    senderInput: _.$('.sender--input'),
+    senderButton: _.$('.sender--button'),
+    letters: _.$A('.letter'),
+    arrow: _.$('.wheel__arrow')
+}
+
+selectors.senderButton.addEventListener('click', () => send(selectors))
+selectors.resultButton.addEventListener('click', () => parse(selectors))
